[naming] Renamed tutorial notebooks (#921)
* renamed notebooks

* propagted renaming to __init__

* fix nblink file name
