--- docs/source/notebooks/Prophet_to_TorchProphet.nblink
@@ -1,3 +1,0 @@
-{
-    "path": "../../../tutorials/feature-use/Prophet_to_TorchProphet.ipynb"
-}
