+++ docs/source/notebooks/prophet_to_torch_prophet.nblink
@@ -1,0 +1,3 @@
+{
+    "path": "../../../tutorials/feature-use/prophet_to_torch_prophet.ipynb"
+}
